AMBARI-124
git-svn-id: https://svn.apache.org/repos/asf/incubator/ambari/trunk@1196984 13f79535-47bb-0310-9956-ffa450edef68

+++ controller/src/main/java/org/apache/ambari/resource/statemachine/ClusterImpl.java
@@ -109,9 +109,8 @@ public class ClusterImpl implements ClusterFSM, EventHandler<ClusterEvent> {
   private ClusterState clusterState;
   private static Log LOG = LogFactory.getLog(ClusterImpl.class);
     
-  public ClusterImpl(Cluster cluster, long revision, 
-      ClusterState clusterState) 
-      throws IOException {
+  public ClusterImpl(Cluster cluster, int revision, 
+      ClusterState clusterState) throws IOException {
     ReadWriteLock readWriteLock = new ReentrantReadWriteLock();
     this.readLock = readWriteLock.readLock();
     this.writeLock = readWriteLock.writeLock();
